requirements.txt
aiohttp
pyrofork
python-dotenv
motor
aiofiles
dnspython
requests
jinja2
tgcrypto-pyrofork